Module libp2p_request_response::codec

Traits

ProtocolName

Types serving as protocol names.

RequestResponseCodec

A RequestResponseCodec defines the request and response types for a RequestResponse protocol or protocol family and how they are encoded / decoded on an I/O stream.