Struct libp2p_request_response::RequestResponse

pub struct RequestResponse<TCodec> where
    TCodec: RequestResponseCodec + Clone + Send + 'static,  { /* fields omitted */ }

A request/response protocol for some message codec.

Implementations

impl<TCodec> RequestResponse<TCodec> where
TCodec: RequestResponseCodec + Clone + Send + 'static,

pub fn new<I>(codec: TCodec, protocols: I, cfg: RequestResponseConfig) -> Self where
I: IntoIterator<Item = (TCodec::Protocol, ProtocolSupport)>,

Creates a new RequestResponse behaviour for the given protocols, codec and configuration.

pub fn send_request(
&mut self,
peer: &PeerId,
request: TCodec::Request
) -> RequestId

Initiates sending a request.

If the targeted peer is currently not connected, a dialing attempt is initiated and the request is sent as soon as a connection is established.

Note: In order for such a dialing attempt to succeed, the RequestResonse protocol must either be embedded in another NetworkBehaviour that provides peer and address discovery, or known addresses of peers must be managed via RequestResponse::add_address and RequestResponse::remove_address.

pub fn send_response(
&mut self,
ch: ResponseChannel<TCodec::Response>,
rs: TCodec::Response
) -> Result<(), TCodec::Response>

Initiates sending a response to an inbound request.

If the ResponseChannel is already closed due to a timeout or the connection being closed, the response is returned as an Err for further handling. Once the response has been successfully sent on the corresponding connection, RequestResponseEvent::ResponseSent is emitted. In all other cases RequestResponseEvent::InboundFailure will be or has been emitted.

The provided ResponseChannel is obtained from an inbound RequestResponseMessage::Request.

pub fn add_address(&mut self, peer: &PeerId, address: Multiaddr)

Adds a known address for a peer that can be used for dialing attempts by the Swarm, i.e. is returned by NetworkBehaviour::addresses_of_peer.

Addresses added in this way are only removed by remove_address.

pub fn remove_address(&mut self, peer: &PeerId, address: &Multiaddr)

Removes an address of a peer previously added via add_address.

pub fn is_connected(&self, peer: &PeerId) -> bool

Checks whether a peer is currently connected.

pub fn is_pending_outbound(&self, peer: &PeerId, request_id: &RequestId) -> bool

Checks whether an outbound request to the peer with the provided PeerId initiated by RequestResponse::send_request is still pending, i.e. waiting for a response.

pub fn is_pending_inbound(&self, peer: &PeerId, request_id: &RequestId) -> bool

Checks whether an inbound request from the peer with the provided PeerId is still pending, i.e. waiting for a response by the local node through RequestResponse::send_response.