Struct libp2p_request_response::throttled::Throttled

pub struct Throttled<C> where
    C: RequestResponseCodec + Send,
    C::Protocol: Sync,  { /* fields omitted */ }

A wrapper around RequestResponse which adds request limits per peer.

Implementations

impl<C> Throttled<C> where
C: RequestResponseCodec + Send + Clone,
C::Protocol: Sync,

pub fn new<I>(c: C, protos: I, cfg: RequestResponseConfig) -> Self where
I: IntoIterator<Item = (C::Protocol, ProtocolSupport)>,
C: Send,
C::Protocol: Sync,

Create a new throttled request-response behaviour.

pub fn from(behaviour: RequestResponse<Codec<C>>) -> Self

Wrap an existing RequestResponse behaviour and apply send/recv limits.

pub fn set_receive_limit(&mut self, limit: NonZeroU16)

Set the global default receive limit per peer.

pub fn override_receive_limit(&mut self, p: &PeerId, limit: NonZeroU16)

Override the receive limit of a single peer.

pub fn remove_override(&mut self, p: &PeerId)

Remove any limit overrides for the given peer.

pub fn can_send(&mut self, p: &PeerId) -> bool

Has the limit of outbound requests been reached for the given peer?

pub fn send_request(
&mut self,
p: &PeerId,
req: C::Request
) -> Result<RequestId, C::Request>

Send a request to a peer.

If the limit of outbound requests has been reached, the request is returned. Sending more outbound requests should only be attempted once Event::ResumeSending has been received from NetworkBehaviour::poll.

pub fn send_response(
&mut self,
ch: ResponseChannel<C::Response>,
res: C::Response
) -> Result<(), C::Response>

Answer an inbound request with a response.

See RequestResponse::send_response for details.

pub fn add_address(&mut self, p: &PeerId, a: Multiaddr)

Add a known peer address.

See RequestResponse::add_address for details.

pub fn remove_address(&mut self, p: &PeerId, a: &Multiaddr)

Remove a previously added peer address.

See RequestResponse::remove_address for details.

pub fn is_connected(&self, p: &PeerId) -> bool

Are we connected to the given peer?

See RequestResponse::is_connected for details.

pub fn is_pending_outbound(&self, p: &PeerId, r: &RequestId) -> bool

Are we waiting for a response to the given request?

See RequestResponse::is_pending_outbound for details.

pub fn is_pending_inbound(&self, p: &PeerId, r: &RequestId) -> bool

Is the remote waiting for the local node to respond to the given request?

See RequestResponse::is_pending_inbound for details.