Type Alias libp2p_request_response::RequestResponseEvent

pub type RequestResponseEvent<TRequest, TResponse> = Event<TRequest, TResponse>;

๐Deprecated since 0.24.0: Use re-exports that omit RequestResponse prefix, i.e. libp2p::request_response::Event

Aliased Type

enum RequestResponseEvent<TRequest, TResponse> {
    Message {
        peer: PeerId,
        message: Message<TRequest, TResponse>,
    },
    OutboundFailure {
        peer: PeerId,
        request_id: RequestId,
        error: OutboundFailure,
    },
    InboundFailure {
        peer: PeerId,
        request_id: RequestId,
        error: InboundFailure,
    },
    ResponseSent {
        peer: PeerId,
        request_id: RequestId,
    },
}

Variants

Message

An incoming message (request or response).

Fields

peer: PeerId

The peer who sent the message.

message: Message<TRequest, TResponse>

The incoming message.

OutboundFailure

An outbound request failed.

Fields

peer: PeerId

The peer to whom the request was sent.

request_id: RequestId

The (local) ID of the failed request.

error: OutboundFailure

The error that occurred.

InboundFailure

An inbound request failed.

Fields

peer: PeerId

The peer from whom the request was received.

request_id: RequestId

The ID of the failed inbound request.

error: InboundFailure

The error that occurred.

ResponseSent

A response to an inbound request has been sent.

When this event is received, the response has been flushed on the underlying transport connection.

Fields

peer: PeerId

The peer to whom the response was sent.

request_id: RequestId

The ID of the inbound request whose response was sent.